+++ test/e2e/storage/nfs_persistent_volume-disruptive.go
@@ -18,6 +18,7 @@ package storage
 
 import (
 	"fmt"
+	"net"
 	"time"
 
 	"github.com/onsi/ginkgo"
@@ -32,6 +33,7 @@ import (
 	e2enode "k8s.io/kubernetes/test/e2e/framework/node"
 	e2epod "k8s.io/kubernetes/test/e2e/framework/pod"
 	e2epv "k8s.io/kubernetes/test/e2e/framework/pv"
+	e2essh "k8s.io/kubernetes/test/e2e/framework/ssh"
 	"k8s.io/kubernetes/test/e2e/framework/volume"
 	"k8s.io/kubernetes/test/e2e/storage/utils"
 )
@@ -40,6 +42,33 @@ type testBody func(c clientset.Interface, f *framework.Framework, clientPod *v1.
 type disruptiveTest struct {
 	testItStmt string
 	runTest    testBody
+}
+
+// checkForControllerManagerHealthy checks that the controller manager does not crash within "duration"
+func checkForControllerManagerHealthy(duration time.Duration) error {
+	var PID string
+	cmd := "pidof kube-controller-manager"
+	for start := time.Now(); time.Since(start) < duration; time.Sleep(5 * time.Second) {
+		result, err := e2essh.SSH(cmd, net.JoinHostPort(framework.GetMasterHost(), sshPort), framework.TestContext.Provider)
+		if err != nil {
+			// We don't necessarily know that it crashed, pipe could just be broken
+			e2essh.LogResult(result)
+			return fmt.Errorf("master unreachable after %v", time.Since(start))
+		} else if result.Code != 0 {
+			e2essh.LogResult(result)
+			return fmt.Errorf("SSH result code not 0. actually: %v after %v", result.Code, time.Since(start))
+		} else if result.Stdout != PID {
+			if PID == "" {
+				PID = result.Stdout
+			} else {
+				//its dead
+				return fmt.Errorf("controller manager crashed, old PID: %s, new PID: %s", PID, result.Stdout)
+			}
+		} else {
+			framework.Logf("kube-controller-manager still healthy after %v", time.Since(start))
+		}
+	}
+	return nil
 }
 
 var _ = utils.SIGDescribe("NFSPersistentVolumes[Disruptive][Flaky]", func() {
@@ -191,7 +220,7 @@ var _ = utils.SIGDescribe("NFSPersistentVolumes[Disruptive][Flaky]", func() {
 
 			ginkgo.By("Observing the kube-controller-manager healthy for at least 2 minutes")
 			// Continue checking for 2 minutes to make sure kube-controller-manager is healthy
-			err = framework.CheckForControllerManagerHealthy(2 * time.Minute)
+			err = checkForControllerManagerHealthy(2 * time.Minute)
 			framework.ExpectNoError(err)
 		})
